Add files via upload
ignore the name (BOB). My brother gave me the idea for BOB. So, this is a multi tool made by Chris. It supports scanning for vulnerabilities, scan for ports, geolocate ips, convert from ip to hostname and the oppside, traceroute, all saved WIFIs (and their passwords), other tools (soem of them need to be installed), scan a wifi, OS, systeminfo, netconn and more! If you have a good anti virus you will need to disable it to be able to install this tool because this tool is for hacking (in a controled enviorment). The tool is made with batch so you can check out the code. Enjoy!!!